requirements.txt
# Core Dependencies (Required)
requests>=2.28.0
beautifulsoup4>=4.11.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
rich>=12.0.0
colorama>=0.4.6
numpy>=1.21.0
lxml>=4.9.0
html5lib>=1.1

# Computer Vision and Video Intelligence (Required for full features)
opencv-python>=4.6.0
Pillow>=9.0.0

# Video Processing (Highly Recommended)
yt-dlp>=2023.1.6

# Optional: Enhanced Features
# Uncomment these for additional capabilities:
# googlesearch-python>=1.2.3  # For Google search (requires API key)
# pandas>=1.5.3  # For data analysis
# scikit-learn>=1.3.0  # For machine learning features
# nltk>=3.8.1  # For natural language processing
# textblob>=0.17.1  # For sentiment analysis
# selenium>=4.15.0  # For advanced web scraping
# webdriver-manager>=4.0.1  # For browser automation
# schedule>=1.2.0  # For scheduled learning sessions
# pytesseract>=0.3.10  # For advanced OCR (requires tesseract)
# librosa>=0.9.0  # For audio processing
# tensorflow>=2.10.0  # For machine learning
